pm-analyzer/backend/requirements.txt
# Core Flask
Flask>=3.0.0
Flask-Cors>=4.0.0
gunicorn>=21.0.0

# AI/ML
google-generativeai>=0.3.0

# Configuration
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0

# HTTP Client
requests>=2.31.0
aiohttp>=3.9.0

# PDF Generation
reportlab>=4.0.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL for BTP
alembic>=1.13.0          # Database migrations

# SAP RFC (optional - install separately if needed)
# pyrfc>=3.0.0

# SAP BTP specific
cfenv>=0.5.3  # Cloud Foundry environment parsing
sap-xssec>=4.0.0  # SAP XSUAA security

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
